pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdl2arch-riscv"
version = "0.1.0"
description = "RISC-V privileged CSR generator built on rdl2arch"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [{ name = "rdl2arch-riscv contributors" }]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "systemrdl-compiler~=1.32",
    "rdl2arch>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "mypy>=1",
]

[project.scripts]
rdl2arch-riscv = "rdl2arch_riscv.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
